Rationale for returning resource in POST responses
Discussed this with Andrey and Leeky. The Microsoft convention is to
return just an ID, the convention of some well-regarded open source APIs
is to return the whole object.

There's no explicit standard around this, and REST doesn't describe what
should happen, beyond saying that the created resource must be
identifiable from the response.